+++ application/backend/tests/unit/runtime/core/components/factories/test_model.py
@@ -7,9 +7,8 @@ from uuid import uuid4
 import pytest
 from instantlearn.utils.constants import SAMModelName
 
-from domain.services.schemas.device import AvailableDeviceSchema
+from domain.services.schemas.device import AvailableDeviceSchema, Device
 from domain.services.schemas.processor import MatcherConfig, PerDinoConfig, SoftMatcherConfig
-from domain.services.schemas.project import Device
 from runtime.core.components.factories.model import DeviceResolver, ModelFactory
 from runtime.core.components.models.passthrough_model import PassThroughModelHandler
